moving paths config into new file
